module6-task1 вынес экшены, декомпозировал на простые функции
