+++ src/reducer.js
@@ -1,4 +1,5 @@
-import {GameType, DefaultGameParams} from "./const";
+import {DefaultGameParams} from "./const";
+import {ActionType} from "./actions";
 import questions from "./mocks/questions";
 
 const initialState = {
@@ -6,38 +7,6 @@ const initialState = {
   maxMistakes: DefaultGameParams.MAX_MISTAKES,
   questions,
   step: DefaultGameParams.STEP
-};
-
-const ActionType = {
-  INCREMENT_MISTAKES: `INCREMENT_MISTAKES`,
-  INCREMENT_STEP: `INCREMENT_STEP`
-};
-
-const isArtistAnswerCorrect = (question, userAnswer) => {
-  return userAnswer.artist === question.song.artist;
-};
-
-const isGenreAnswerCorrect = (question, userAnswer) => {
-  return userAnswer.every((it, i) => {
-    return it === (question.answers[i].genre === question.genre);
-  });
-};
-
-const ActionCreator = {
-  incrementStep: () => ({
-    type: ActionType.INCREMENT_STEP,
-    payload: 1
-  }),
-  incrementMistake: (question, userAnswer) => {
-    const answerIsCorrect = question.type === GameType.ARTIST
-      ? isArtistAnswerCorrect(question, userAnswer)
-      : isGenreAnswerCorrect(question, userAnswer);
-
-    return {
-      type: ActionType.INCREMENT_MISTAKES,
-      payload: answerIsCorrect ? 0 : 1
-    };
-  }
 };
 
 const reducer = (state = initialState, action) => {
@@ -60,4 +29,4 @@ const reducer = (state = initialState, action) => {
   return state;
 };
 
-export {reducer, ActionType, ActionCreator};
+export {reducer};
