+++ src/reducer.test.js
@@ -1,5 +1,5 @@
-import {ActionCreator, ActionType, reducer} from "./reducer.js";
-
+import {reducer} from "./reducer.js";
+import {incrementStep, incrementMistake, ActionType} from "../../actions";
 
 const questions = [{}];
 
@@ -9,7 +9,7 @@ describe(`reducer test`, () => {
       step: -1,
       mistakes: 0,
       questions,
-    }, ActionCreator.incrementStep())).toEqual({
+    }, incrementStep())).toEqual({
       step: 0,
       mistakes: 0,
       questions,
@@ -19,14 +19,14 @@ describe(`reducer test`, () => {
 
 describe(`Action creator test`, () => {
   it(`Action creator for incrementing step returns correct action`, () => {
-    expect(ActionCreator.incrementStep()).toEqual({
+    expect(incrementStep()).toEqual({
       type: ActionType.INCREMENT_STEP,
       payload: 1,
     });
   });
 
   it(`Action creator for incrementing mistake returns action with 0 payload if answer for artist is correct`, () => {
-    expect(ActionCreator.incrementMistake({
+    expect(incrementMistake({
       type: `artist`,
       song: {
         artist: `correct`,
@@ -54,7 +54,7 @@ describe(`Action creator test`, () => {
   });
 
   it(`Action creator for incrementing mistake returns action with 1 payload if answer for artist is incorrect`, () => {
-    expect(ActionCreator.incrementMistake({
+    expect(incrementMistake({
       type: `artist`,
       song: {
         artist: `correct`,
@@ -82,7 +82,7 @@ describe(`Action creator test`, () => {
   });
 
   it(`Action creator for incrementing mistake returns action with 0 payload if answer for genre is correct`, () => {
-    expect(ActionCreator.incrementMistake({
+    expect(incrementMistake({
       type: `genre`,
       genre: `jazz`,
       answers: [
@@ -107,7 +107,7 @@ describe(`Action creator test`, () => {
   });
 
   it(`Action creator for incrementing mistake returns action with 1 payload if answer for genre is incorrect`, () => {
-    expect(ActionCreator.incrementMistake({
+    expect(incrementMistake({
       type: `genre`,
       genre: `jazz`,
       answers: [
